fix(frontend): enhance dashboard sections
