+++ packages/frontend/src/components/dashboard/index.tsx
@@ -18,7 +18,7 @@ import { PinnedWorkflows } from "./widgets/PinnedWorkflows";
 import { QuickActions } from "./widgets/QuickActions";
 import { RecentActivity } from "./widgets/RecentActivity";
 import { RecentRuns } from "./widgets/RecentRuns";
-import { UpcomingSchedule } from "./widgets/UpcomingSchedule";
+import { UpcomingScheduleTimeline } from "./widgets/UpcomingScheduleTimeline";
 
 export const Dashboard = () => {
   const { t } = useTranslate();
@@ -26,54 +26,61 @@ export const Dashboard = () => {
   return (
     <Grid container gap={3} flexDirection="column">
       <PageHeader icon={Home} title={t("title.dashboard")} />
-      
-      {/* 1. Key Metrics (Top Priority) */}
-      <Grid container spacing={3}>
-         <Grid size={12}>
-            <KPICards />
-         </Grid>
+
+      <Grid size={12}>
+        <KPICards />
       </Grid>
 
       {/* 2. Main Dashboard Layout */}
       <Grid container spacing={3}>
         {/* Left Column: Activity & Actions */}
-        <Grid size={{ md: 8, xs: 12 }} container spacing={3} flexDirection="column">
-             {/* Quick Actions Bar */}
-             <Grid size={12}>
-                <QuickActions />
-             </Grid>
+        <Grid
+          size={{ md: 8, xs: 12 }}
+          container
+          spacing={3}
+          flexDirection="column"
+        >
+          {/* Quick Actions Bar */}
+          <Grid size={12}>
+            <QuickActions />
+          </Grid>
 
-             <Grid size={12}>
-                <PinnedWorkflows />
-             </Grid>
+          <Grid size={12}>
+            <PinnedWorkflows />
+          </Grid>
 
-             <Grid size={12}>
-                <RecentRuns />
-             </Grid>
+          <Grid size={12}>
+            <RecentRuns />
+          </Grid>
 
-            <Grid container spacing={3}>
-                <Grid size={{ md: 6, xs: 12 }}>
-                    <ConversationSnapshot />
-                </Grid>
-                <Grid size={{ md: 6, xs: 12 }}>
-                    <RecentActivity />
-                </Grid>
+          <Grid container spacing={3}>
+            <Grid size={{ md: 6, xs: 12 }}>
+              <ConversationSnapshot />
             </Grid>
+            <Grid size={{ md: 6, xs: 12 }}>
+              <RecentActivity />
+            </Grid>
+          </Grid>
         </Grid>
 
         {/* Right Column: Status & Schedule */}
-        <Grid size={{ md: 4, xs: 12 }} container spacing={3} flexDirection="column">
-            <Grid size={12}>
-                <AttentionRequired />
-            </Grid>
+        <Grid
+          size={{ md: 4, xs: 12 }}
+          container
+          spacing={3}
+          flexDirection="column"
+        >
+          <Grid size={12}>
+            <AttentionRequired />
+          </Grid>
 
-            <Grid size={12}>
-                <UpcomingSchedule />
-            </Grid>
+          <Grid size={12}>
+            <UpcomingScheduleTimeline />
+          </Grid>
 
-            <Grid size={12}>
-                <IntegrationsHealth />
-            </Grid>
+          <Grid size={12}>
+            <IntegrationsHealth />
+          </Grid>
         </Grid>
       </Grid>
     </Grid>
